@frankmcsherry 2017-11-23
For example, once candidate reconciliation would be to write a method

```rust
fn embalm<T: Abomonation>(bytes: &mut [u8])
```

which interprets `bytes` as a `&T` and performs all of the appropriate address sanitization, retaining that ability for people who are excited about this, but making it be part of a second pass for people who are less excited about it.