@frankmcsherry 2017-11-24
I think the ability to put in `embalm` afterwards is a good solution. I'm less certain I know how to sanitize things with all the new "stash discriminators in special fields" logic going on, so happy to have that demoted as a core feature.